dispatch: support transposing shaders during output
This is done by flipping the vertex coordinates both in the vertex list
and (again) in the vertex shader. The net result of this is that shaders
are rendered to the correct rect but the mapping of all other vertex
attributes is transposed with respect to the vertex coordinates.

Currently non-functional, but used to implement rotation in the
following commit.