meson: refactor the config.h mess
I have no idea why I had such a complicated mess instead of this way of
doing it, which seems vastly simpler, and works by simply installing it
into the correct path to begin with.

I think it was because, in some distant past, I installed the entire
headers subdir, and therefore could not have a stray config.h.in or
meson.build inside there. Or something. My memory is hazy.

Anyway, simplify the whole config.h build thing.