shaders/icc: improve 3DLUT generation performance
Using TYPE_RGBA_FLT, even as output, is extremely slow. It's much faster
to just generate it UINT format and perform the conversion manually.

It would be really nice if we could support e.g. UNORM textures in
sh_lut, to avoid even this last conversion step. But oh well.