dispatch: expose pl_dispatch_reset_frame
The old header erroneously claimed that this function is not needed
publicly. However, it is, to advance the state of the PRNG (and perform
pass garbage collection).

Otherwise, users of the `pl_dispatch` API have no way to e.g. deband
textures properly - they will always get the same PRNG seed.