vulkan: purge pass layout optimization
The recent CI failure on master revealed a deeper issue that can occur
when re-using the same shader between different types of textures.

From some quick benchmarking I can tell absolutely no performance
difference here anyways, and in practice almost all of these formats are
basically equivalent as far as the driver is concerned, so we really,
really shouldn't bother with this sort of brittle optimization.