vulkan: support timeline semaphores
Needed for compatibility with upstream FFmpeg/lavfi changes, which
switched to timeline semaphores internally (and as such, we need to do
the same to support wrapping vulkan hwframes).

This requires a rather annoying API break, where pl_vulkan_hold and
pl_vulkan_release have to be extended by a uint64_t value to support
timeline semaphores. Just do the needful and break the type completely,
introducing a new wrapping struct for the semaphore/value pair.

Also try and auto-load the feature, even though nothing technically
depends on it yet, as a convenience to the user. (And because we will
soon require it ourselves)