vulkan: kill off vk_signal and use timeline semaphores
vk_signal was always a bit of a clunky state-tracking application that
required a complicated mess of unsignalling logic, semaphore pooling,
and generally led to unpredictable behavior (especially given that
semaphores were not clearly associated with individual GPU objects but
rather re-used ad hoc).

This commit also fixes a few bugs en passant, for example the current
lack of a pipeline barrier for same-queue render passes sans events
(which are *not* implicitly synchronized, without an external subpass
dependency, which we don't use).

The new vk_sem abstraction went through several iterations, but is
currently split into read and write accesses, to try and emit the
minimum possible barriers. (A much simpler design would be to treat all
accesses as readwrite, but I felt like overengingeering this bit. The
numbers just barely justify it, if you squint.)

So, overall, not only is this vastly more correct, it's also a bit more
performant. Murdering the entire bird nest with one big stone.