vulkan: fix re-entrant callback processing
Recursively polling commands from a callback, while commands are already
being polled, results in callbacks for the same command being run twice
(possibly exhausting the stack if that callback itself again decides to
process callbacks a third time, etc.).

The fix for this is twofold:

1. Make sure we dequeue commands to be reset *before* running its
   callbacks, to make sure re-entrant calls to vk_poll_commands don't
   see the same command again.

2. To avoid unexpected callback order as much as possible, keep track of
   remaining callbacks for the same command and flush them from the
   re-entrant call before processing the next command. This isn't
   terribly inefficient because we can just pointer tricks instead of
   memory management.

Obviously, doing this sort of thing fundamentally breaks FIFO ordering -
but if a callback *requires* some sort of polling command (e.g.
pl_buf_read) to be executed *while* inside the callback, then breaking
FIFO order is fundamentally unavoidable - obviously, the command is
expecting to see the results of callbacks that come after it in FIFO
order.

So basically, this current solution means that a re-entrant callback
will see *its* processing delayed until the callbacks it depends on are
processed, but any other callback is unaffected.

Incidentally, this is recursively re-entrant, only hitting an eventual
stack size limit. The use of a single `pending_callbacks` pointer does
not harm us here, because every re-entrant callback will just spin up
another stack frame of `flush_callbacks`, even if the dependency
hierarchy is several layers deep.