Revert "dispatch: always upgrade to compute shaders"
This reverts commit 58cd6f183f2ffd213a418ceb1155fa29def60cbb.

It's a performance regression for others, and in general even for me the
gains are only really manifest when using exclusively compute queues. It
probably also has something to do with the fact that RADV still doesn't
have a transfer queue, so upgrading to compute shaders allows avoiding
cross-queue dependencies, or something.

I'm not sure. But at any rate, this does represent a regression for
sorts for `vo_gpu_next` in particular.