vulkan: add scary warning on pl_vulkan_release
I just realized that it's never quite clear what the lifetime
requirements of this semaphore are, since the user has no way of knowing
what libplacebo's internal command queue state is.

In practice this is unlikely to matter because the only real users of
this API hold on to their semaphores until the entire GPU instance has
been destroyed. But potential other users should be wary of this
pitfall nonetheless.