vulkan/malloc: avoid edge case for extreme alignment
If the alignment is very high, higher than the requirement for the slab
size alignment, the current heuristic logic can lead to large amounts of
wasted memory. So apply the alignment before the heuristic.

In order to help debug this sort of error case better, also print the
alignment on OOM.