shaders/sampling: mark pl_shader_deband non-resizable
Resizing the output of a debanding shader is generally a very bad idea,
because it incurs an extra performance hit for no real benefit. To
discourage users from doing this accidentally, mark the debanding shader
as non-resizable - effectively fixing the intended debanding resolution
at invocation time.

This effectively also fixes an issue where enabling debanding in the
renderer was significantly slower than intended, when not combining it
with another scaler.

Not bumping the API version for this because the public signature is
unchanged.