renderer: make pl_render_image_mix work on single frames
This can still be disabled by using `skip_caching_single_frames` to
revert back to the old logic. That said, I decided to default it to
enabled, effectively making `pl_render_image_mix` always go through the
mixer cache. This is a tiny performance regression for content with the
exact same FPS as the display, but a massive performance gain for 24 fps
content, as well as a near-infinite performance gain while paused.

Users that specifically *know* they will never re-draw certain frames
are rare (and basically only includes mpv), so I'm comfortable forcing
them to manually advertise this knowledge.

In theory, we could also have auto-detection for this inside the
`pl_render_image_mix` code itself, since it can generate a rough
estimate of the source fps and vsync duration to figure out when this
exceptional case arises, but I decided not to do that in the current
commit. May re-evaluate in the future.