vulkan: fix partial rendering with !load_target
We always need to pick LOAD_OP_LOAD for partial renders, even if we
don't care about blending, because this is required to preserve
framebuffer contents outside the scissor area.

We could in theory get around this in some cases, when we know we're
rendering to partial crops that perfectly align with the render area
granularity, but we'd need to know this ahead of time, which we
explicitly can't (by the fact that it's fixed in advantage).

Bite the bullet and always load the framebuffer contents. In theory it
might be faster to just use a partial/misaligned render area instead,
but I have no clue how to benchmark that as my hardware does not make a
distinction whatsoever.

Closes #111