shaders/colorspace: clean up pl_shader_color_map
This was written a bit weirdly. Make the logic more straightforward.

Not technically identical, since a < was replaced by a == in the new
version, but that edge case is fairly obscure and the harm done is
miniscule. (It only possibly results in a redundant application of gamma
curve, and only if linear scaling is explicitly disabled)