renderer: mix in the frame's gamma function
The previous change in 28612113 brought an unfortunate regression
alongside it, which was to always force gamma function application
during frame mixing. This was not the intent - the intent was to
preserve previous linearization (and, as a hidden upside, also allow
mixing in linear light).

Brings back the dreaded `fidx` assert. While I could reasonably work
around them, I've decided that I still really don't think that `fidx`
should ever be 0 (this would, at best, result in a black frame). I'd
rather get to the bottom of any circumstance in which we hit that case.

Fixes 286121137116df94729892d982fc066b7699c126