gpu: better document semantics of `cached_program`
It was brought to my attention that this can essentially be used as a
vector to perform undefined behavior (and therefore possible code
execution?) by passing maliciously crafted cached programs.

In practice this is not a huge concern because these cache objects tend
to come from ~/.cache and such - and if you're writing to that, you can
already write to ~/.zshrc and take control over the machine anyways. But
it should absolutely be documented.