vulkan/command: nuke command queue
With the switch to timeline semaphores, and the resulting absence of the
unsignaling mechanism, we no longer have any reason to hold on for
commands longer than necessary. As such, we can immediately submit all
commands to the GPU, rather than delaying such submission until the user
flushes queued commands explicitly.

This completely gets rid of the command queue, the command object
associations (which were only needed for partial flushing), the last_cmd
state tracking and the relevant API functions.

No performance difference (as expected), but vastly simpler code.