utils/frame_queue: fix thread race
If another thread pushes a frame while the current thread is reading
from a `mix` object, the pl_frame pointers can be invalidated due to
reallocation of `p->queue`.

Fix this by making `p->queue` an array of pointers, rather than an array
of frame structs, and allocating all entries on the heap. Ugly but
works.