shaders: tweak LUT selection logic
I have a use case where using texture is actually preferable to a
hard-coded LUT, making me strongly reconsider whether we want to use
hard-coded LUTs for even 256-sized LUTs. Split it up into a soft and
hard limit, so we can make the soft limit as low as we want.