shaders/lut: support scientific notation numbers
I think the spec was ambiguous over whether or not these are valid, but
cube files in the wild use them, so we need to support them, too.

Requires both accepting a (redundant) + prefix on integers, and changing
the digit list in shaders/lut.

Fixes https://github.com/mpv-player/mpv/issues/9637