d3d11/tex: use the specified format and array index for RTVs and UAVs
Unlike mpv's D3D11 RA, which uses a separate function for wrapping
swapchain surfaces and video textures, libplacebo uses the same function
for both. This means that, depending on the texture's bind flags, it
might try to create an RTV or UAV for a wrapped video texture or texture
array. For video textures, that might not work unless the format is
specified and for array textures, the array index should be specified,
so specify both.