utils/libav: don't auto-map dovi metadata
This is actually a regression for vf_libplacebo. We need to handle this
in a better way. Naively auto-applying it is nice, but causes too many
surprises.

I will think up something different. In the meantime, avoid breaking the
status quo.

Not bothering to bump the API version for this because it's a very
recent change and doesn't affect the public API.