shaders/colorspace: clamp luma in PL_GAMUT_DESATURATE
The assumption that gamut mapping will always be the output of
appropriately tone-mapped content is false. The most obvious
counter-example is when mixing SDR overlays onto HDR content that has
already been tone-mapped. If the overlay has any sort of ringing (e.g.
as a result of scaling), it will contain out-of-range luma values.