vulkan/swapchain: make start_frame/submit_frame more robust
These always have to be submitted in lockstep, and doing something like
e.g. resizing the swapchain in the meantime leads to undefined behavior.
Safeguard this with a combination of changes:

1. Hold on to a lock for the duration of the call, thus preventing the
   fbo state from getting invalidated while being rendered to.
2. Make `last_imgidx` contain -1 while not rendering a frame, so we can
   assert on it being set to something valid.

Note that this also fixes a subtle bug where we accidentally passed the
address of `p->last_imgidx` (an int) to pImageIndices (uint32_t *).