colorspace: fix color repr normalization
When e.g. normalizing 10-bits-in-16 color values, the resulting color
values are as they would be for a 16-bit sample, not for a 10-bit
sample. So we need to set the color depth to the sample depth, not the
other way around.

Otherwise, this breaks YCbCr decoding later on, resulting in a green
shift (among other issues).

Should fix https://github.com/mpv-player/mpv/issues/9403