d3d11/context: check for debug layer presence before enabling it
Previously, we used the DXGI_ERROR_SDK_COMPONENT_MISSING status code to
determine the if the debug layer failed to load, however that didn't
work with Direct3D 11.0, because it returns E_FAIL instead. A more
reliable method, which is used in Microsoft sample code, is to determine
the presence of the debug layer before the actual device creation by
attempting to create a device with the NULL driver and the debug layer
enabled.

This change also raises the log level of the message printed when the
debug layer is missing, since it should be easy to tell if a lack of
debug messages is because the debug flag was disabled or the debug layer
was not installed correctly.