d3d11/formats: only set gatherable when sampleable is set
libplacebo requires formats to be sampleable if they are gatherable.
With the real D3D11 runtime, that isn't a problem, since FL11_1 has no
gatherable UINT or SINT formats. However, DXVK seems to expose
gatherable UINT and SINT formats. Since we never consider integer
formats to be sampleable, we unset the sampleable flag (but not the
gatherable flag) which made libplacebo assert when verifying the format
list (in gpu.c.) Instead, only set gatherable when sampleable is set.