utils/upload: drop unnecessary asserts
These used to be relevant in the past, but past refactors have gradually
loosened up these restrictions to the point where it's no longer
necessary to check anything at all. Even in the case that it would be
necessary, the `pl_tex_upload` call itself would be responsible for
appropriate error checking, so asserting here is always wrong.