shaders/colorspace: make gamut mapping defaults more conservative
PL_GAMUT_DARKEN had a lot of undesirable side effects. It's not a good
way of fixing this problem, and definitely not something we should hoist
on users by default if "HDR is too dark" is already the persistent
complaint around HDR.

Clipping is probably the most conservative default. It doesn't introduce
the weird white-ish hue shifts that people complain about with regards
to DESATURATE. It's also the closest to what users expect from naive
implementations.

Not bumping the API version for this because it doesn't change the
API/ABI, even though it affects the public header.