shaders/sampling: expand polar sampling fast path
Instead of calculating wbase by hard-coding a dependency on the pos_map
function, communicate the appropriate `base` from the per-pixel
calculated values of the correct shader invocation.

Compute the correct shader invocation based whether or not the src rect
is flipped, thus allowing this shader to work properly with flipped
rects.

Removes both of the current fast path restrictions, allowing the compute
shader to work with flipped rects and even the sampler2D interface.