shaders/sampling: fix oversample shader
No idea what was wrong with this, but it wasn't producing the expected
results at all. Rewrote it from the ground up to be correct, before
simplifying the math to get it as close as possible to the previous
state again.

Can probably be simplified further, but this doesn't need to be
super-optimized anyway.

Fixes https://code.videolan.org/videolan/libplacebo/-/issues/185