renderer: correctly initialize neutral chroma
This should not be exactly 0.5, but a value slightly higher than 0.5
(depending on the image bit depth).

It's possible that we should go with the mpv approach of just setting
the chroma channels to 0 altogether for grayscale formats. But that
would require some sort of API change, which I'm unwilling to do as part
of this commit.

The 0.5 constant for floating point YCbCr is almost surely not actually
correct, but that preserves the status quo. cf. #155

Fixes https://github.com/mpv-player/mpv/issues/9969