gpu: move shader logging to the gpu.c wrappers
I wanted to change the way this is handled such that shaders also get
printed when being re-loaded from the cache. Rather than re-adding the
same code to all three GPU instances, I decided it makes more sense to
just move it to the common helper code.

This does unfortunately re-order most shader debug messages such that
the errors get printed _above_ the re-print of the source code (on
errors), but this was the status quo anyways for vulkan. I can live with
it.

Closes https://code.videolan.org/videolan/libplacebo/-/issues/192