d3d11/swapchain: make initial color config logging less confusing
d3d11_sw_colorspace_hint is utilized both for runtime hints as well
as the initial swap chain configuration to enable consistency
between received swap chains and newly created ones.

Thus, add an internal function which contains the argument of whether
this call was internal or not. In the colorspace hint call, it is
set to false, and during swap chain initialization it is set to true.
Thus when hints are passed, logging will talk about received hints
and when the function is called during initialization, logging will
talk about initial configuration.