vulkan: stop using implicit layout transitions
This makes the vk_tex_barrier code unnecessarily buggy. In particular,
the way the code is currently written, we omit pipeline barriers here
*anyway*. And trying to reason about when the synchronization scope can
be ignored and when not gave me a headache so just get rid of this
implicit layout transition business and use the same pipeline barrier
that we use for everything else.

The overhead is negligible anyway.