vulkan: hold on to coherent buffers while synchronizing
Even for a coherent buffer, we technically have no guaranteed thet
'vkCmdPipelineBarrier' has completed yet until the command callbacks are
fired. To solve this issue, take the reference unconditionally, and
instead move the coherency conditional to `invalidate_buf`.

As an aside, checking for `buf_vk->mem.data` here is redundant, as this
code will only ever be hit in any of the host readable/writable
conditions.