renderer: add fallback for absurd frame mixes
Sometimes a frame mix simply contains very sparse collection of frames
that are far too far in the past and the future to actually get mixed.
In this case, fall back to ZOH semantics.

Do this by retrying the loop in order to enable the caching logic to
work. Not the prettiest work-around but simple and effective.

Also add a test case to prevent further regressions of this case.

Fixes https://code.videolan.org/videolan/libplacebo/-/issues/202