shaders/icc: reimplement custom gamma estimation method
This avoids a dependency on upstream cmsDetectRGBProfileGamma. It's
slightly less precise, using only 256 sample points instead of 4096, but
this difference barely seems to matter in practice, as this is only a
rough estimate anyway.

The main motivation of this change is to avoid a dependency on the
relatively high lcms2 2.13, but it also brings with it the other upside
of allowing us to tune the algorithm to our needs (e.g. not filtering
out linear sections, because we care more about estimating the overall
response to minimize LUT error, and less about finding the technical
gamma).