vulkan/utils_gen: also check datadir for vk.xml
This is useful in cases where users build the vulkan-headers (and thus
place vk.xml) inside their custom prefixes, in which case it makes sense
to look there as well.

Do this with highest priority - if a user is building against custom
vulkan headers, we probably want to use them. (They might be pinned to
known-good versions)