shaders/icc: drop unnecessary and confusing flags
As a result of b9fcede8, we now use this profile for a whole lot more
than merely transforming a color cube. As such, specifying the grid
points to 2 would be a major disaster. The only reason it didn't blow up
was because `cmsFLAGS_NOOPTIMIZE` actually overrides it.

Drop the unnecessary, confusing, and potentially harmful flags.