vulkan: implement array_size_constants
MSL function constants, the MoltenVK equivalent of push constants,
do not support this.

In theory, we only need to disable this on MoltenVK, but we can use
VK_KHR_portability_subset as a proxy for this case. (And it's probably
problematic on other platforms as well, so a good idea to avoid in
general)