shaders/sampling: reduce polar compute recompilations
This change reduces recompilation overhead on GPUs that don't support
array push constants, as well as when using dynamic_constants.

Rounding up to multiples of 8 is a good value to pick because it aligns
well with typical radii and upscaling ratios. Notably, an exact 2x
integer upscale with default ewa_lanczos results in a natural size of
24x24 on 1024-thread devices,