meson: explicitly test for glad2 version
Building fails if the user only has glad1 installed, but not glad2, with
this error: argument --api: Invalid api-string: "gl:core,gles2,egl"

And if the user has neither glad1 nor glad2, it just fails with the
unhelpful "module not found". Improve both by explicitly testing for
this in the build system.