vulkan/swapchain: rewrite/fix swapchain recreation logic
So, from the vulkan spec, swapchains are put into a retired state
whether the call to `vkCreateSwapchainKHR` succeeds or fails, so we
always need to garbage collect them.

In addition, the point about multiple "stale" swapchains being forbidden
is also bogus, the vulkan spec explicitly clarifies that this is
allowed.

Finally, this solves a number of theoretical race conditions that were
present as a result of this struct not being locked with respect to
callbacks, by passing the (wrapped) handle directly.