renderer: make PL_LUT_CONVERSION override ICC profiles
This was the intended behavior, and documented as such in mpv, but not
actually documented nor implemented in libplacebo.

Reshuffle the logic around a bit to implement things this way. Notably,
move color blindness simulation to the source space, which should
generally be a bit faster anyway because it skips linearization in many
cases.