shaders/icc: support forced black point linearization
This overrides the built-in BPC logic and progressively mixes a linear
ramp into the near-black entries of the generated LUT. May help fix
"bad" monitor profiles, containing invalid colorimetric tables.

Disabled by default because it effectively disables the 3DLUT near
black, which may not be desired, especially for well-behaved profiles.

Closes: https://code.videolan.org/videolan/libplacebo/-/issues/219