renderer: add support for deinterlacing
Doing deinterlacing directly in the renderer is really the only sane way
to do it, due to the following advantages:

- keeps all params in the same place
- natural frame invalidation when deinterlacing params change
- natural integration into the pl_renderer_save/load shader cache
- integration into the render stats / timings
- only deinterlaces when a frame is actually needed/used
- possibility of merging with other shader passes

Annoyingly, as a result of API limitations, we currently have to do
deinterlacing before plane merging, which is a needless slowdown. But
fixing that is a problem for another day.