meson/vulkan: refactor `vulkan-link`
Instead of a bool. This is an unfortunate config API break, but
hopefully not a too drastic one as most people will have left this on
the default setting. Even so, give it a new name to disambiguate it from
the old setting.

This change is a functional ipmrovement as it allows us to distinguish
between 'auto' and 'enabled' for 'vulkan-link', up from the previous
default of being effectively equivalent to 'auto'. I also wanted to
simplify the code to distinguish more clearly between the headers and
loader.

A minor benefit of the variable renaming is that the value printed in
the configuration summary now matches the option name.