colorspace: ignore PQ minimum luminance
Typical PQ sources out there in the wild have mastering display minimum
luminances significantly above the black point they are actually
mastered to (typically true black). SImply ignore the tagged minimum
luminance completely for PQ spaces.

We have to do this for both the source *and* target because that ensures
that blindly copying metadata from the source to the target continues
being a no-op.

This does unfortunately break our ability to influence the true black
point of PQ tone mapped signals, but given that PQ monitors in practice
seem to effectively tone map the low end anyway, this is probably
intended.

Fixes: https://github.com/mpv-player/mpv/issues/10592