renderer: refactor pl_render_info
With the addition of error diffusion, as well as (technically) the
addition of YCbCr subsampled outputs, we can actually have multiple
shader passes even in the blend stage.

The old hack of re-using the `index` field to indicate the number of
frames being blended is simply inadequate. Fix it properly by adding a
new field.