utils/libav: test for <vulkan.h> dynamically
Now that we bundle <vulkan.h>, PL_HAVE_VULKAN is no longer a "reliable"
method of detecting whether vulkan is enabled. To be fair, it never was,
but this change revealed this bug.

The real issue here is that <vulkan.h> cannot be unconditionally assumed
to be available, so we need to test for it somehow. Re-introduce the
dynamic check (works on GCC/clang) and leave a scary note hoping to get
MSVC users to include <vulkan.h> manually.