renderer: perform debanding first
Note that this commit on its own is a performance regression because it
disables the optimization where luma debanding can get merged with the
color conversation pass. This shortcoming be addressed by the following
commits. It is also a different regression in the case of cropped
rendering, whereas previously only the cropped region was debanded, now
the entire frame is. This regression stands unmodified.

That said, this is a massive improvement in the case where any sort of
luma prescaler touches the luma plane, because it means we perform
debanding before luma scaling in this case.

Fixes: https://code.videolan.org/videolan/libplacebo/-/issues/230