renderer: delay plane shader execution
Instead of executing all shaders right away, delay them to be executed
on-demand, and solve the error handling issue by deferring error
handling to a later stage.

This is not strictly speaking functional on its own, but it sets the
stage for a planned change to enable more liberal pass merging when
scaling is not required.

Also rewrite some of the logic to make it clearer what's going on, and
avoid referencing the confusing `ref->img.w/h` variables for the wrong
purpose.