shaders/deinterlacing: fix differences to C yadif
These two changes mysteriously appeared in the CUDA version of Yadif,
from which this shader was implemented. Undo them to bring the code in
line with the C version again.

I'm really, really annoyed by having to hard-code assumptions about the
video bit depth here. But I don't feel like this is adequate
justification to break API just to plumb it through. Besides, the visual
difference appears to be insignificant.