shaders/deinterlacing: fix field calculation
This logic was copied blindly from the CUDA version, but comparing it
aginst the C yadif made me conclude that it's wrong. (As the naming of
the variable clearly indicates, `is_second_field` should be true when
field is *not* equal to first)

Fix the logic (and get rid of the basically redundant variable, "field
== first_field" is clearly understandable, so why confuse matters by
inverting the boolean name arbitrarily?).