shaders/sampling: refactor pl_shader_sample_ortho
Investigating a bug with broken yuv422 scaling on a cropped texture made
me conclude that this API is fundamentally broken. We cannot discard the
unused component from inside this function without knowing what the
source crop was. We need it to be less magic and to move responsibility
onto the API user.

Since this is a significant break in API, rename it to
`pl_shader_sample_ortho2` and preserve the original name for backwards
compatibility. Annoying, but necessary, because otherwise this regresses
current users of `pl_shader_sample_ortho` (e.g. vs-placebo).