shaders/custom: remove stray whitespace
No need to parse this off the start beacuse we strip the line
immediately afterwards anyway. Change it for consistency with the other
parsers. There's an argument to be had on whether or not this is
intended behavior, as it allows mis-parsing something like
"//!DESCRIPTION" as "//!DESC RIPTION", but oh well.