colorspace: add ACES ultra-wide gamuts
Some very wide or unusual ICC profiles can currently fail to be handled
because we have no primary set large enough to contain their primaries.
This commit adds AP0 and AP1. AP0 is defined to encompass the entire Yxy
gamut, and AP1 is the widest possible strictly positive subset of this.