renderer: properly scale ortho first pass
The current logic still failed when confronted with subpixel offsets.
There's an easy and a hard way to fix this properly - the hard way is to
round the dimensions to explicit integer pixels and then adjust the src2
rect to compensate.

The easy way is to just completely skip scaling in the first direction
by brute-forcing this to the texture dimensions.

I'll choose the easy way.