tone_mapping: add pl_tone_map_params_infer
Resolving `pl_tone_map_params` at the time of `pl_tone_map_generate` is
too late, because the latter can be generated dynamically based on the
detected HDR peak, while we generally want to preserve a consistent
'look & feel' even when using HDR peak detection.

To fix this logic early on without duplicating logic, we need to expose
the logic as a function that the tone mapping shader can call into.