renderer: fix memory leak of pass.tmp
Remove unnecessry allocation of `render_pass.tmp`.

When commit [0] added this reallocation of render_pass.tmp for reuse of
the struct it was still necessary to do so.
Down the line `pass_init()` was introduced which also allocates the
field tmp.
The object presiously pointed to by this field was not freed however,
leading to a memory leak of 32 bytes everytime the reallocation was
executed.

Remove the explicit allocation to fix the leak.

[0] eeed2d7f3999 ("renderer: fix several bugs related to pl_render_image reuse")