gpu: optimize pl_tex_blit_compute for 1-pixel blits
This is a special case that we use to emulate texture clears on planar
vulkan textures, so it's worth skipping the source coordinate
calculation in this case. This has the effect of making the texture load
not data-dependent, which in theory means it can be shared for the
entire work-group.