vulkan: implement planar synchronization
Since `pl_tex_export` is deprecated, we only need to worry about
`pl_vulkan_hold/release`. To ensure we always have a consistent state,
the planar texture itself does not track any state - all synchronization
happens at the per-plane level.

This requires some legwork to ensure we transition all planes as part of
a single command (with a single semaphore submission), but it's not too
bad.