renderer: correctly re-linearize un-linearized linear inputs
As the comment mentions, if facing PL_COLOR_TRC_LINEAR input, we decide
to un-linearize for scaling purposes. We should undo this logic when it
comes time to output to screen.

In theory, another way of accomplishing this would be to change the
signature of `pl_shader_color_map` to allow it to take an override for
the input TRC. I think I would prefer this version, but I don't know how
to break the API in a graceful way.

I might fix this properly at a later point in time.